consolidating config package
