--- internal/config/credentials.go
@@ -1,9 +1,0 @@
-package config
-
-type CredentialConfiguration struct {
-	Username   string
-	Password   string
-	PassLM     string
-	PassNT     string
-	PassNTLMv2 string
-}
