--- internal/config/proxy.go
@@ -1,10 +1,0 @@
-package config
-
-type ProxyConfiguration struct {
-	Domain       string
-	ProxyAddress []string
-	NoProxy      []string
-	Address      string
-	Port         []int
-	Credentials  CredentialConfiguration
-}
